+++ src/media/audio/pcm/mod.rs
@@ -7,6 +7,7 @@
 mod tests;
 
 mod buf; // PcmBuf
+mod layout; // PcmLayout
 mod sample; // PcmSample, PcmSampleType
 mod spec; // PcmSpec
 
@@ -14,6 +15,7 @@ crate::structural_mods! { // _mods
     _mods {
         pub use super::{
             buf::*,
+            layout::*,
             sample::*,
             spec::*,
         };